+++ src/compojure/api/methods.clj
@@ -1,0 +1,3 @@
+(ns compojure.api.methods)
+
+(def all-methods #{:get :head :patch :delete :options :post :put})